+++ src/main/java/fr/univ_lille/gitlab/classrooms/assignments/AssignmentServiceImpl.java
@@ -1,15 +1,12 @@
 package fr.univ_lille.gitlab.classrooms.assignments;
 
-import fr.univ_lille.gitlab.classrooms.domain.*;
-import fr.univ_lille.gitlab.classrooms.gitlab.GitlabApiFactory;
+import fr.univ_lille.gitlab.classrooms.classrooms.Classroom;
+import fr.univ_lille.gitlab.classrooms.classrooms.ClassroomService;
+import fr.univ_lille.gitlab.classrooms.gitlab.Gitlab;
 import fr.univ_lille.gitlab.classrooms.quiz.QuizService;
 import fr.univ_lille.gitlab.classrooms.users.ClassroomUser;
 import jakarta.transaction.Transactional;
-import org.gitlab4j.api.GitLabApi;
 import org.gitlab4j.api.GitLabApiException;
-import org.gitlab4j.api.models.AccessLevel;
-import org.gitlab4j.api.models.GroupParams;
-import org.gitlab4j.api.models.Project;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -23,23 +20,20 @@ class AssignmentServiceImpl implements AssignmentService {
 
     private final QuizService quizService;
 
-    private final GitLabApi gitLabApi;
-
-    private final GitlabApiFactory gitlabApiFactory;
+    private final Gitlab gitlab;
 
     private final ClassroomService classroomService;
 
     private final AssignmentRepository assignmentRepository;
 
-    private final StudentExerciseRepository studentExerciseRepository;
+    private final StudentAssignmentRepository studentAssignmentRepository;
 
-    AssignmentServiceImpl(QuizService quizService, GitLabApi gitLabApi, GitlabApiFactory gitlabApiFactory, ClassroomService classroomService, AssignmentRepository assignmentRepository, StudentExerciseRepository studentExerciseRepository) {
+    AssignmentServiceImpl(QuizService quizService, Gitlab gitlab, ClassroomService classroomService, AssignmentRepository assignmentRepository, StudentAssignmentRepository studentAssignmentRepository) {
         this.quizService = quizService;
-        this.gitLabApi = gitLabApi;
-        this.gitlabApiFactory = gitlabApiFactory;
+        this.gitlab = gitlab;
         this.classroomService = classroomService;
         this.assignmentRepository = assignmentRepository;
-        this.studentExerciseRepository = studentExerciseRepository;
+        this.studentAssignmentRepository = studentAssignmentRepository;
     }
 
     @Override
@@ -51,41 +45,32 @@ class AssignmentServiceImpl implements AssignmentService {
     @Transactional
     public void acceptAssigment(Assignment assignment, ClassroomUser student) throws GitLabApiException {
         assignment.accept(student);
+        this.assignmentRepository.save(assignment);
 
         if (assignment instanceof ExerciseAssignment exerciseAssignment) {
-            // get a gitlab api client ith the teacher's rights
-            var teacherGitlabApi = this.gitlabApiFactory.userGitlabApi(assignment.getClassroom().getTeacher());
+            // create the project in gitlab
+            var project = gitlab.createProject(exerciseAssignment, student);
 
-            var templateRepo = exerciseAssignment.getGitlabRepositoryTemplateId();
-            if (templateRepo == null || templateRepo.isBlank()) {
-                // get student id in gitlab
-                var studentUserId = teacherGitlabApi.getUserApi().getUser(student.getName()).getId();
-
-                // create a blank project
-                var projectParams = new Project()
-                        .withName(assignment.getName() + "-" + student.getName())
-                        .withNamespaceId(exerciseAssignment.getGitlabGroupId());
-                var project = teacherGitlabApi.getProjectApi().createProject(projectParams);
-
-                // grant the student access to its project
-                teacherGitlabApi.getProjectApi().addMember(project.getId(), studentUserId, AccessLevel.MAINTAINER);
-
-                // create the student exercise in the database
-                var studentExercise = new StudentExercise();
-                studentExercise.setAssignment(exerciseAssignment);
-                studentExercise.setStudent(student);
-                studentExercise.setGitlabProjectId(project.getId());
-                studentExercise.setGitlabProjectUrl(project.getWebUrl());
-                this.studentExerciseRepository.save(studentExercise);
-            }
+            // create the student exercise assignment
+            var studentExercise = new StudentExerciseAssignment();
+            studentExercise.setAssignment(exerciseAssignment);
+            studentExercise.setStudent(student);
+            studentExercise.setGitlabProjectId(project.getId());
+            studentExercise.setGitlabProjectUrl(project.getWebUrl());
+            this.studentAssignmentRepository.save(studentExercise);
         }
-
-        this.assignmentRepository.save(assignment);
+        else if (assignment instanceof QuizAssignment quizAssignment) {
+            // create the student quiz assignment
+            var studentExercise = new StudentQuizAssignment();
+            studentExercise.setAssignment(quizAssignment);
+            studentExercise.setStudent(student);
+            this.studentAssignmentRepository.save(studentExercise);
+        }
     }
 
     @Override
-    public List<StudentExercise> getAssignmentResults(Assignment assignment) {
-        return this.studentExerciseRepository.findAllByAssignment(assignment);
+    public List<StudentAssignment> getAssignmentResults(Assignment assignment) {
+        return this.studentAssignmentRepository.findAllByAssignment(assignment);
     }
 
     @Override
@@ -108,18 +93,11 @@ class AssignmentServiceImpl implements AssignmentService {
     @Override
     @Transactional
     public Assignment createExerciseAssignment(Classroom classroom, String assignmentName, String repositoryId) throws GitLabApiException {
-        var groupPath = assignmentName.trim().replaceAll("[^\\w\\-.]", "_");
-        var groupParams = new GroupParams()
-                .withName(assignmentName)
-                .withPath(groupPath)
-                .withDescription("Gitlab group for the assignment " + assignmentName)
-                .withParentId(classroom.getGitlabGroupId());
-        var group = this.gitLabApi.getGroupApi().createGroup(groupParams);
-
         var exerciseAssignment = new ExerciseAssignment();
         exerciseAssignment.setName(assignmentName);
         exerciseAssignment.setGitlabRepositoryTemplateId(repositoryId);
-        exerciseAssignment.setGitlabGroupId(group.getId());
+
+        this.gitlab.createGroup(exerciseAssignment, classroom);
 
         classroom.addAssignment(exerciseAssignment);
 
@@ -130,7 +108,12 @@ class AssignmentServiceImpl implements AssignmentService {
     }
 
     @Override
-    public StudentExercise getAssignmentResultsForStudent(Assignment assignment, ClassroomUser student) {
-        return this.studentExerciseRepository.findByAssignmentAndStudent(assignment, student);
+    public StudentAssignment getAssignmentResultsForStudent(Assignment assignment, ClassroomUser student) {
+        return this.studentAssignmentRepository.findByAssignmentAndStudent(assignment, student);
+    }
+
+    @Override
+    public List<StudentAssignment> getAllStudentAssignmentsForAClassroom(Classroom classroom, ClassroomUser student) {
+        return this.studentAssignmentRepository.findByAssignmentClassroomAndStudent(classroom, student);
     }
 }