+++ src/main/java/fr/univ_lille/gitlab/classrooms/classrooms/ClassroomController.java
@@ -1,9 +1,9 @@
-package fr.univ_lille.gitlab.classrooms.domain;
+package fr.univ_lille.gitlab.classrooms.classrooms;
 
+import fr.univ_lille.gitlab.classrooms.gitlab.Gitlab;
 import fr.univ_lille.gitlab.classrooms.users.ClassroomUser;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.servlet.http.HttpSession;
-import org.gitlab4j.api.GitLabApi;
 import org.gitlab4j.api.GitLabApiException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
@@ -18,20 +18,22 @@ import java.util.UUID;
 @RolesAllowed("TEACHER")
 class ClassroomController {
 
+    private static final String CLASSROOM_MODEL_ATTRIBUTE = "classroom";
+
     private final ClassroomService classroomService;
 
-    private final GitLabApi gitLabApi;
+    private final Gitlab gitlab;
 
     private static final System.Logger LOGGER = System.getLogger(ClassroomController.class.getName());
 
-    public ClassroomController(ClassroomService classroomService, GitLabApi gitLabApi) {
+    public ClassroomController(ClassroomService classroomService, Gitlab gitlab) {
         this.classroomService = classroomService;
-        this.gitLabApi = gitLabApi;
+        this.gitlab = gitlab;
     }
 
     @GetMapping("/new")
     String newClassroom(Model model) throws GitLabApiException {
-        model.addAttribute("groups", this.gitLabApi.getGroupApi().getGroups());
+        model.addAttribute("groups", this.gitlab.getGroupsOfConnectedUser());
         return "classrooms/new";
     }
 
@@ -44,17 +46,14 @@ class ClassroomController {
     @GetMapping("/{classroomId}")
     String showClassroom(@PathVariable UUID classroomId, Model model) {
         var classroom = this.classroomService.getClassroom(classroomId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        model.addAttribute(CLASSROOM_MODEL_ATTRIBUTE, classroom);
 
         try {
-            var group = gitLabApi.getGroupApi().getGroup(classroom.getGitlabGroupId());
-            model.addAttribute("gitlabGroupUrl", group.getWebUrl());
+            model.addAttribute("gitlabGroupUrl", gitlab.getGroupURI(classroom));
         }
         catch (GitLabApiException e){
             LOGGER.log(System.Logger.Level.ERROR, e.getMessage());
         }
-
-
-        model.addAttribute("classroom", classroom);
 
         return "classrooms/view";
     }
@@ -63,15 +62,16 @@ class ClassroomController {
     @RolesAllowed({"TEACHER", "STUDENT"})
     String showJoinClassroom(@PathVariable UUID classroomId, Model model) {
         var classroom = this.classroomService.getClassroom(classroomId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        model.addAttribute(CLASSROOM_MODEL_ATTRIBUTE, classroom);
 
-        model.addAttribute("classroom", classroom);
         return "classrooms/join";
     }
 
     @PostMapping("/{classroomId}/join")
     @RolesAllowed({"TEACHER", "STUDENT"})
-    String joinClassroom(@PathVariable UUID classroomId, @ModelAttribute("user") ClassroomUser student, Model model, HttpSession session) throws GitLabApiException {
+    String joinClassroom(@PathVariable UUID classroomId, @ModelAttribute("user") ClassroomUser student, Model model, HttpSession session) {
         var classroom = this.classroomService.getClassroom(classroomId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        model.addAttribute(CLASSROOM_MODEL_ATTRIBUTE, classroom);
 
         this.classroomService.joinClassroom(classroom, student);
 
@@ -81,7 +81,6 @@ class ClassroomController {
             return "redirect:" + redirect;
         }
 
-        model.addAttribute("classroom", classroom);
         return "classrooms/joined";
     }