+++ src/main/java/fr/univ_lille/gitlab/classrooms/classrooms/ClassroomService.java
@@ -1,4 +1,4 @@
-package fr.univ_lille.gitlab.classrooms.domain;
+package fr.univ_lille.gitlab.classrooms.classrooms;
 
 import fr.univ_lille.gitlab.classrooms.users.ClassroomUser;
 import org.gitlab4j.api.GitLabApiException;
@@ -11,9 +11,16 @@ public interface ClassroomService {
 
     List<Classroom> getAllClassrooms();
 
+    /**
+     * Returns all the classrooms the student has joined
+     * @param student the student
+     * @return a list of classrooms
+     */
+    List<Classroom> getAllJoinedClassrooms(ClassroomUser student);
+
     Optional<Classroom> getClassroom(UUID uuid);
 
-    void joinClassroom(Classroom classroom, ClassroomUser student) throws GitLabApiException;
+    void joinClassroom(Classroom classroom, ClassroomUser student);
 
     void createClassroom(String classroomName, Long parentGitlabGroupId, ClassroomUser teacher) throws GitLabApiException;