+++ src/main/resources/db/migration/V16__CLASSROOM_USER_GITLAB_ID.sql
@@ -1,0 +1,1 @@
+alter table classroom_user add gitlab_user_id bigint;