+++ src/test/java/fr/univ_lille/gitlab/classrooms/classrooms/ClassroomControllerMVCTest.java
@@ -1,13 +1,12 @@
-package fr.univ_lille.gitlab.classrooms.domain;
+package fr.univ_lille.gitlab.classrooms.classrooms;
 
+import fr.univ_lille.gitlab.classrooms.gitlab.Gitlab;
 import fr.univ_lille.gitlab.classrooms.users.ClassroomUser;
 import fr.univ_lille.gitlab.classrooms.users.WithMockStudent;
 import fr.univ_lille.gitlab.classrooms.users.WithMockTeacher;
-import org.gitlab4j.api.GitLabApi;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.mockito.Answers;
 import org.mockito.ArgumentCaptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -39,10 +38,10 @@ class ClassroomControllerMVCTest {
     @MockBean
     private ClassroomService classroomService;
 
-    @MockBean(answer = Answers.RETURNS_DEEP_STUBS)
-    private GitLabApi gitLabApi;
+    @MockBean
+    private Gitlab gitlab;
 
-    private UUID classroomId = UUID.randomUUID();
+    private final UUID classroomId = UUID.randomUUID();
 
     @BeforeEach
     void setUp() {
@@ -81,6 +80,27 @@ class ClassroomControllerMVCTest {
                     .andDo(print())
                     .andExpect(status().isForbidden());
         }
+    }
+
+    @Test
+    @WithMockTeacher
+    void newClassroom_shouldListTheGroupsOfTheTeacher() throws Exception {
+        mockMvc.perform(get("/classrooms/new"))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        verify(this.gitlab).getGroupsOfConnectedUser();
+    }
+
+    @Test
+    @WithMockTeacher
+    void showClassroom_shouldListClassroomAssignments() throws Exception {
+        mockMvc.perform(get("/classrooms/"+classroomId))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        verify(this.classroomService).getClassroom(classroomId);
+        verify(this.gitlab).getGroupURI(any());
     }
 
     @Test